Cancel previously queued jobs
We trigger workflow jobs in E2E tests. If the E2E tests exit
unexpectedly due to reasons like a hard timeout, the triggered workflow
jobs remain in the queue. When a new run of E2E tests provisions some
runners, they pick up the old jobs, causing the newly triggered jobs to
get stuck in the queue. We should keep only the latest job and cancel
the rest.